Fixing translation variance of expressions in forms.
Up until now, we apparently have not tested translation variant expressions in forms. For affine domain transformations (aka no blending) the implementation falsely assumed that the entire form is translation invariant. But this is of course not true. It only applies to the Jacobian.

The downside of this simple fix is that I guess it will produce unused variables. But let's see.