Changes behaviour of generate_all_hyteg_forms.py
The commit changes the behaviour of generate_all_hyteg_forms.py with respect
to the selection of forms to generate. It introduces a new function

   assemble_list_of_forms_to_generate()

that extracts the list of forms to generate from the list of defined forms
based on the command-line options in the following fashion:

* if neither --geometry, nor --filter is given, all defined forms are
  selected

* if only --filter is given, it is applied to the list of all defined forms,
  which are generated for each geometry they support

* if only --geometry is given, it extracts all forms that support the given
  geometry element type

* if both --filter and --geometry are given, the filter gets applied first;
  then the function removes all froms that do not support the selected
  geometry

The commit touches on MR !48 and issue #45.

As a spin-off the commit makes the FormInfo class sortable.