Fix generation of epsilon and full_stokes forms
Generation of theses forms did no longer work due to problems with the update
from a scalar to a vectorial treatment of trial and test spaces.

The commit fixes this, by making form_func() convert trial and test to
TensorialVectorFunctionSpace internally.

Note: I am not too happy with the changes needed to statisfy mypy.
      I was not able to keep the differentiation between TestSpace
      and TrialSpace for the arguments that are of type
      TensorialVectorFunctionSpace. Ideas welcome.